Add, modularize supersedes/superseded_by display
